+++ CHANGELOG.md
@@ -1,7 +1,20 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [Version 1.3.1] — 2020-09-17
+
+### Fixed
+
+* Prevent the order count transient from being set before $wc_order_types is populated ([#52])
+
+### Changed
+
+* Move the limiter initialization from "woocommerce_loaded" to "init" ([#52])
+* Change the format of the `{current_interval}` and `{next_interval}` placeholders if the interval is less than 24 hours ([#53])
+
 
 ## [Version 1.3.0] — 2020-07-16
 
@@ -85,6 +98,7 @@ Initial plugin release.
 [Version 1.2.0]: https://github.com/nexcess/limit-orders/releases/tag/v1.2.0
 [Version 1.2.1]: https://github.com/nexcess/limit-orders/releases/tag/v1.2.1
 [Version 1.3.0]: https://github.com/nexcess/limit-orders/releases/tag/v1.3.0
+[Version 1.3.1]: https://github.com/nexcess/limit-orders/releases/tag/v1.3.1
 [#5]: https://github.com/nexcess/limit-orders/pull/5
 [#6]: https://github.com/nexcess/limit-orders/pull/6
 [#8]: https://github.com/nexcess/limit-orders/pull/8
@@ -100,3 +114,5 @@ Initial plugin release.
 [#41]: https://github.com/nexcess/limit-orders/pull/41
 [#42]: https://github.com/nexcess/limit-orders/pull/42
 [#43]: https://github.com/nexcess/limit-orders/pull/43
+[#52]: https://github.com/nexcess/limit-orders/pull/52
+[#53]: https://github.com/nexcess/limit-orders/pull/53